removed unnecessary pom sections, bundle properties, and System.out. Added source code url, renamed packages.
