+++ .gitignore
@@ -1,3 +1,2 @@
 .DS_Store
 **/target/**
-/target/
