+++ modules/SimilarityComputer/src/img/weights-ok.png
