--- modules/SimilarityComputer/src/main/resources/Parsers/Bundle.properties
@@ -1,41 +1,0 @@
-OpenIDE-Module-Display-Category=Import
-OpenIDE-Module-Long-Description=\
-    <h2>Convert Excel and csv files to networks (including dynamic!)\
-    </h2>\n\nThis plugin helps you import Excel files and csv files into Gephi, by transforming them into networks.\
-    <br>\nIt takes the rows of your file and let you define which relations should be found in it.\
-    <br>\n<br>\nPlease post issues and ask for feature request on <a href="https://github.com/seinecle/My-Plugins-for-Gephi">Github</a>.\
-    <br>\n<br>\nOr contact me (Clement Levallois) via Twitter @seinecle.\
-    <br>\n<br>\n\n\n<h3>Release History:</h3>\n<b>Version 1.4.2 (September 2014)</b><br>\n-> Bug fix for numerical values in Excel file: now "96" is accepted as a valid name for the name of node (before, the line would have been ignored). \
-    Added a warning that for Excel, only .xlsx files are accepted (not .xls)<br>\n<br>\n<b>Version 1.4.1 (September 2014)</b><br>\n-> Bug fix for time intervals in the Excel file: now "2013-05-12, 2013-05-13" is accepted. \
-    The bug was that the space after the comma would have caused the line to be ignored.\
-    <br>\n<br>\n<b>Version 1.4 (September 2014)</b><br>\n-> Added support for temporal networks.\
-    <br>\n<br>\n<b>Version 1.3 (March 2014)</b><br>\n-> Bug fix (suggested by Wouters A.H. \
-    Spekkink)<br>\nSame names in different columns of the Excel of csv files are now correctly analyzed.\
-    <br>\n<br>\n<b>Version 1.0 (October 2013)</b><br>\n-> Initial release.\
-    <br>
-OpenIDE-Module-Name=GephiSimilarityComputer
-OpenIDE-Module-Short-Description=Converts Excel files and csv files into networks
-Panel1.jButtonSelectFile.text=select file
-Panel3.jLabelSecondConnectedAgent.text=jLabel2
-Panel3.jLabelFirstConnectedAgent.text=jLabel1
-Panel3.jTextArea1.text=Example:\nLet's imagine you selected "authors" for agents in the previous screen. Suppose that the field for "authors" is made of several co-authors, separated by semicolon.\n=> Please choose "semicolon" as a delimiter.\n(leave the selection empty if no delimiter applies).
-Panel3.jLabelSecondAgentAlso.text=Is this field made of subfields? If so, what is the delimiter?
-Panel3.jLabel3.text=Is this field made of subfields? If so, what is the delimiter?
-Panel4.jCheckBoxInnerLinks.text=create links between jLabel3 themselves, and between jLabel5 themselves
-Panel4.jLabel2.text=A network will be created, where:
-Panel4.jLabel1.text=The conversion process will start when you click the "Finish" button.
-Panel4.jLabel3.text=jLabel3
-Panel4.jLabel4.text=are connected to
-Panel4.jLabel5.text=jLabel5
-Panel4.jLabel6.text=- The edge weight will represent the number of connections between any two agents
-Panel4.jLabel7.text=- Each node (agent) will have an attribute (called "frequency") representing how many times this agent was present in the file.
-Panel2.jLabel5.text=Example: for a network of co-authors, pick the field of co-authors twice!
-Panel2.jLabel3.text=is connected to
-Panel2.jLabel2.text=1.  This type of agent:
-Panel2.jLabel1.text=What are the connections made of?
-Panel2.jLabel4.text=2. This type of agent:
-Panel1.jCheckBoxHeadersIncluded.text=file includes headers (column titles)
-Panel1.jLabel3.text=Select the Excel sheet containing the data:
-Panel1.jLabel2.text=default text delimiter is a double quote --> " <--  You can choose another one:
-Panel1.jLabel1.text=Choose the field delimiter:
-Panel1.jTextFieldTextDelimiter.text=text delimiter
