--- modules/SimilarityComputer/src/main/resources/Wizard/Bundle.properties
@@ -1,28 +1,0 @@
-OpenIDE-Module-Name=GephiSimilarityComputer
-OpenIDE-Module-Short-Description=Sort nodes alphabetically
-Panel1.jButtonSelectFile.text=select file
-Panel3.jLabelSecondConnectedAgent.text=jLabel2
-Panel3.jLabelFirstConnectedAgent.text=jLabel1
-Panel3.jTextArea1.text=Example:\nLet's imagine you selected "authors" for agents in the previous screen. Suppose that the field for "authors" is made of several co-authors, separated by semicolon.\n=> Please choose "semicolon" as a delimiter.\n(leave the selection empty if no delimiter applies).
-Panel3.jLabelSecondAgentAlso.text=Is this field made of subfields? If so, what is the delimiter?
-Panel3.jLabel3.text=Is this field made of subfields? If so, what is the delimiter?
-Panel2.jLabel5.text=Example: for a network of co-authors, pick the field of co-authors twice!
-Panel2.jLabel3.text=is connected to
-Panel2.jLabel2.text=1.  This type of agent:
-Panel2.jLabel1.text=What are the connections made of?
-Panel2.jLabel4.text=2. This type of agent:
-Panel1.jCheckBoxHeadersIncluded.text=file includes headers (column titles)
-Panel1.jLabel2.text=default text delimiter is a double quote --> " <--  You can choose another one:
-Panel1.jTextFieldTextDelimiter.text=
-Panel1.jLabelFieldDelimiter.text=Choose the field delimiter:
-Panel4.jLabel3.text=If you want to create a dynamic network, where is the field for time?
-Panel4.jTextArea1.text=Let's imagine you each of your data entry has a time stamp. You could create a network changing in time, where nodes and their connections appear and disappear along a timeline.\n2 time formats allowed: 1984, 2014-09-17 (but not 2014-17-09).\nDates (1984) but also durations (1984,1986) are allowed in this field.  For duration, use the comma to separate the start and end dates.
-Panel4.jLabel1.text=Leave blank if the network is not dynamic
-Panel4.jLabel2.text=! Works only for Excel files at the moment (not csv!).
-Panel1.jLabelExcelSheets.text=Select the Excel sheet containing the data:
-Panel1.jTextArea1.text=Your file should contain:\n\n- first col: the name of the entities which will be the nodes in the network\n- second column: an attribute of this entity\n- third column: a second attribute of this entity\n- fourth column: etc...\n
-Panel3.jLabel6.text=- The edge weight will represent the strength of the similarity between two entities
-Panel3.jLabel1.text=The conversion process will start when you click the "Finish" button.
-Panel2.jCheckBoxWeight.text=Is there a value attached to each feature, to indicate its strength or intensity?
-Panel1.jCheckBoxMenuItem1.text=jCheckBoxMenuItem1
-Panel2.jLabelPicture.text_1=
