+++ modules/SimilarityComputer/src/main/resources/net/clementlevallois/controller/Bundle.properties
@@ -1,0 +1,10 @@
+OpenIDE-Module-Display-Category=Import
+OpenIDE-Module-Long-Description=\
+    <h2>Create a network of entities based on their similarities</h2>\n\nThis plugin helps you convert a list of entities and their features into a network of entities, based on how similar they are in terms of features<br>\nPlease post issues and ask for feature request on <a href="https://github.com/seinecle/gephi-plugins/tree/similaritycomputer">Github</a>.\
+    <br>\n<br>\nOr contact me (Clement Levallois) via Twitter @seinecle.\
+    <br>\n<br>\n\n\n<h3>Release History:</h3>\n\
+    <b>Version 1.1 (Feb 2017)</b><br>\n-> Release for Gephi 0.9.1 with additional features.\
+    <b>Version 1.0 (September 2014)</b><br>\n-> Initial release.\
+    <br>
+OpenIDE-Module-Name=Similarity Computer
+OpenIDE-Module-Short-Description=Creates a network between entities, based on their similarities.
