+++ modules/SimilarityComputer/src/main/resources/net/clementlevallois/wizard/Bundle.properties
@@ -1,0 +1,29 @@
+OpenIDE-Module-Display-Category=Import
+OpenIDE-Module-Long-Description=\
+    <h2>Create a network of entities based on their similarities</h2>\n\nThis plugin helps you convert a list of entities and their features into a network of entities, based on how similar they are in terms of features<br>\nPlease post issues and ask for feature request on <a href="https://github.com/seinecle/gephi-plugins/tree/similaritycomputer">Github</a>.\
+    <br>\n<br>\nOr contact me (Clement Levallois) via Twitter @seinecle.\
+    <br>\n<br>\n\n\n<h3>Release History:</h3>\n\
+    <b>Version 1.1 (Feb 2017)</b><br>\n-> Release for Gephi 0.9.1 with additional features.\
+    <b>Version 1.0 (September 2014)</b><br>\n-> Initial release.\
+    <br>
+OpenIDE-Module-Name=Similarity Computer
+OpenIDE-Module-Short-Description=Creates a network between entities, based on their similarities.
+#Panel 1
+Panel1.jTextFieldTextDelimiter.text=
+Panel1.jCheckBoxMenuItem1.text=jCheckBoxMenuItem1
+Panel1.jCheckBoxHeadersIncluded.text=file includes headers (column titles)
+Panel1.jLabel2.text=default text delimiter is a double quote --> " <--  You can choose another one:
+Panel1.jLabelFieldDelimiter.text=Choose the field delimiter:
+Panel1.jTextArea1.text=Your file should contain:\n- first col: the name of the entities which will be the nodes in the network\n - second column: an attribute of this entity\n- third column: a second attribute of this entity\n- fourth column: etc... 
+Panel1.jButtonSelectFile.text=select file
+Panel1.jLabelExcelSheets.text=Select the Excel sheet containing the data:
+#
+#Panel 2
+#
+Panel2.jLabelPicture.text_1=
+Panel2.jCheckBoxWeight.text=Is there a value attached to each feature, to indicate its strength or intensity?
+#
+#Panel 3
+#
+Panel3.jLabel6.text=- The edge weight will represent the strength of the similarity between two entities
+Panel3.jLabel1.text=The conversion process will start when you click the "Finish" button.
